refactor: fix major audit issues — DRY, remove legacy CLI, update CHANGELOG
Address major findings from strict repository audit:

1. README: remove hardcoded test count (was stale at 4870)
2. DRY: consolidate _has_valid_agent_result() from 3 copies into canonical
   implementation in agent_runner.py; consolidate _has_valid_judge_result()
   from 2 copies into judge_runner.py
3. Remove legacy scylla Click CLI (scylla/cli/main.py) — all experiment
   management uses scripts/manage_experiment.py
4. Move progress.py from scylla/cli/ to scylla/e2e/ (used by orchestrator)
5. CHANGELOG: add [0.1.0] version history and [Unreleased] changes
6. Update doc consistency hook to accept README without hardcoded test count
