+++ pixi.lock
@@ -3882,7 +3882,7 @@ packages:
 - pypi: ./
   name: scylla
   version: 0.1.0
-  sha256: 5812021ede749b06b4a46ac3a74c2c6d1d2cce9f20ce9406e8193c3b1988126c
+  sha256: e752c3e4782b8f7b7249d23d6bc4f16c911d0609828a7dad45a3651f2b17792a
   requires_dist:
   - click>=8.0,<9
   - pydantic>=2.0,<3
