+++ pyproject.toml
@@ -30,9 +30,6 @@ dependencies = [
     "pyyaml>=6.0,<7",
 ]
 
-[project.scripts]
-scylla = "scylla.cli.main:main"
-
 [project.optional-dependencies]
 dev = [
     "pytest>=7.0,<10",
