--- scylla/cli/__init__.py
@@ -1,25 +1,0 @@
-"""CLI module for ProjectScylla.
-
-This module provides command-line interface and progress display.
-"""
-
-from scylla.cli.progress import (
-    EvalProgress,
-    ProgressDisplay,
-    RunProgress,
-    RunStatus,
-    TierProgress,
-    format_duration,
-    format_progress_bar,
-)
-
-__all__ = [
-    "EvalProgress",
-    # Progress
-    "ProgressDisplay",
-    "RunProgress",
-    "RunStatus",
-    "TierProgress",
-    "format_duration",
-    "format_progress_bar",
-]
