--- tests/unit/cli/__init__.py
@@ -1,1 +1,0 @@
-"""CLI module tests."""
